remove all last cache code from schema module
