+++ schema-engine-tag/src/main/java/org/apache/iotdb/db/metadata/tagSchemaRegion/utils/ShowTimeSeriesResultUtils.java
@@ -44,7 +44,7 @@ public class ShowTimeSeriesResultUtils {
         schemaEntry.getTSDataType(),
         schemaEntry.getTSEncoding(),
         schemaEntry.getCompressionType(),
-        schemaEntry.getLastTime(),
+        Long.MAX_VALUE,
         new HashMap<>(),
         new HashMap<>(),
         null,
@@ -68,7 +68,7 @@ public class ShowTimeSeriesResultUtils {
         schemaEntry.getTSDataType(),
         schemaEntry.getTSEncoding(),
         schemaEntry.getCompressionType(),
-        schemaEntry.getLastTime(),
+        Long.MAX_VALUE,
         new HashMap<>(),
         new HashMap<>(),
         null,
