+++ server/src/main/java/org/apache/iotdb/db/metadata/idtable/IDTable.java
@@ -34,7 +34,6 @@ import org.apache.iotdb.db.metadata.idtable.entry.SchemaEntry;
 import org.apache.iotdb.db.metadata.idtable.entry.TimeseriesID;
 import org.apache.iotdb.db.metadata.plan.schemaregion.write.ICreateAlignedTimeSeriesPlan;
 import org.apache.iotdb.db.metadata.plan.schemaregion.write.ICreateTimeSeriesPlan;
-import org.apache.iotdb.tsfile.read.TimeValuePair;
 import org.apache.iotdb.tsfile.utils.Pair;
 import org.apache.iotdb.tsfile.write.schema.IMeasurementSchema;
 
@@ -86,30 +85,6 @@ public interface IDTable {
    * @throws MetadataException if insert plan's aligned value is inconsistent with device
    */
   // IDeviceID getSeriesSchemas(InsertPlan plan) throws MetadataException;
-
-  /**
-   * get last cache of the timeseies
-   *
-   * @param timeseriesID timeseries ID of the timeseries
-   * @throws MetadataException if the timeseries is not exits
-   */
-  TimeValuePair getLastCache(TimeseriesID timeseriesID) throws MetadataException;
-
-  /**
-   * update last cache of the timeseies
-   *
-   * @param timeseriesID timeseries ID of the timeseries
-   * @param pair last time value pair
-   * @param highPriorityUpdate is high priority update
-   * @param latestFlushedTime last flushed time
-   * @throws MetadataException if the timeseries is not exits
-   */
-  void updateLastCache(
-      TimeseriesID timeseriesID,
-      TimeValuePair pair,
-      boolean highPriorityUpdate,
-      Long latestFlushedTime)
-      throws MetadataException;
 
   /** clear id table and close file */
   void clear() throws IOException;
