+++ server/src/main/java/org/apache/iotdb/db/metadata/mnode/EntityMNode.java
@@ -18,8 +18,6 @@
  */
 package org.apache.iotdb.db.metadata.mnode;
 
-import org.apache.iotdb.db.metadata.lastCache.container.ILastCacheContainer;
-import org.apache.iotdb.db.metadata.lastCache.container.LastCacheContainer;
 import org.apache.iotdb.db.metadata.mnode.visitor.MNodeVisitor;
 
 import java.util.Collections;
@@ -37,8 +35,6 @@ public class EntityMNode extends InternalMNode implements IEntityMNode {
   private transient volatile Map<String, IMeasurementMNode> aliasChildren = null;
 
   private volatile boolean isAligned = false;
-
-  private volatile Map<String, ILastCacheContainer> lastCacheMap = null;
 
   @Override
   public String getFullPath() {
@@ -160,26 +156,6 @@ public class EntityMNode extends InternalMNode implements IEntityMNode {
   @Override
   public void setAligned(boolean isAligned) {
     this.isAligned = isAligned;
-  }
-
-  public ILastCacheContainer getLastCacheContainer(String measurementId) {
-    checkLastCacheMap();
-    return lastCacheMap.computeIfAbsent(measurementId, k -> new LastCacheContainer());
-  }
-
-  @Override
-  public Map<String, ILastCacheContainer> getTemplateLastCaches() {
-    return lastCacheMap == null ? Collections.emptyMap() : lastCacheMap;
-  }
-
-  private void checkLastCacheMap() {
-    if (lastCacheMap == null) {
-      synchronized (this) {
-        if (lastCacheMap == null) {
-          lastCacheMap = new ConcurrentHashMap<>();
-        }
-      }
-    }
   }
 
   @Override
