+++ server/src/main/java/org/apache/iotdb/db/metadata/mnode/IEntityMNode.java
@@ -18,8 +18,6 @@
  */
 package org.apache.iotdb.db.metadata.mnode;
 
-import org.apache.iotdb.db.metadata.lastCache.container.ILastCacheContainer;
-
 import java.util.Map;
 
 public interface IEntityMNode extends IMNode {
@@ -48,8 +46,4 @@ public interface IEntityMNode extends IMNode {
   boolean isAligned();
 
   void setAligned(boolean isAligned);
-
-  ILastCacheContainer getLastCacheContainer(String measurementId);
-
-  Map<String, ILastCacheContainer> getTemplateLastCaches();
 }
