+++ server/src/main/java/org/apache/iotdb/db/metadata/mnode/IMeasurementMNode.java
@@ -19,7 +19,6 @@
 package org.apache.iotdb.db.metadata.mnode;
 
 import org.apache.iotdb.commons.path.MeasurementPath;
-import org.apache.iotdb.db.metadata.lastCache.container.ILastCacheContainer;
 import org.apache.iotdb.tsfile.file.metadata.enums.TSDataType;
 import org.apache.iotdb.tsfile.write.schema.IMeasurementSchema;
 
@@ -42,10 +41,6 @@ public interface IMeasurementMNode extends IMNode {
   long getOffset();
 
   void setOffset(long offset);
-
-  ILastCacheContainer getLastCacheContainer();
-
-  void setLastCacheContainer(ILastCacheContainer lastCacheContainer);
 
   boolean isPreDeleted();
